requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.3.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==8.3.0
numba==0.58.1
numerize==0.12
numpy==1.24.4
packaging==23.2
pandas==2.0.3
patsy==0.5.6
pillow==10.2.0
pkgutil-resolve-name==1.3.10
plotly==5.20.0
plotly-express==0.4.1
protobuf==4.25.3
pyarrow==15.0.1
pydeck==0.8.1b0
pygments==2.17.2
pynndescent==0.5.11
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
statsmodels==0.14.1
streamlit==1.32.2
streamlit-option-menu==0.3.12
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
watchdog==4.0.0
zipp==3.18.1
